upload code

+++ src/java/com/poly/controller/MenuController.java
@@ -32,8 +32,15 @@ public class MenuController {
     @Autowired
     MenuService menuService;
 
+    AccountController accController = new AccountController();
+    
     @RequestMapping()
     public String initiate(ModelMap model, HttpSession session) {
+        if (Objects.equals(menuService.checkLogin(session), Boolean.FALSE)) {
+            model.addAttribute(ConstantManager.ERROR_POPUP, ConstantManager.NO_ACCEPTED_LOGIN);
+            return accController.initiate(model, session);
+        }
+        model.put("link", "menu.htm");
         List<Menu> m = menuService.getFather();
         model.put("menuList", m);
         return "menu";
@@ -41,9 +48,13 @@ public class MenuController {
 
     @RequestMapping(params = "getSubMenu")
     public String getSubMenu(ModelMap model, HttpSession session, @RequestParam("id") String id) {
-        List<Menu> m = menuService.getSon(id);
-        model.put("subMenu", m);
-        model.put("menuName", menuService.getMenuById(id).getName());
+        if (Objects.equals(menuService.checkLogin(session), Boolean.FALSE)) {
+            model.addAttribute(ConstantManager.ERROR_POPUP, ConstantManager.NO_ACCEPTED_LOGIN);
+            return accController.initiate(model, session);
+        }
+        model.put("subMenu", menuService.getSon(id));
+        model.put("menuFatherName", menuService.getMenuById(id).getName());
+        model.put("menuFatherList", menuService.getFather());
         model.put("menuFatherId", id);
         return "subMenu";
     }
@@ -51,6 +62,10 @@ public class MenuController {
     @RequestMapping(params = "insertMenu", method = RequestMethod.POST)
     public String insert(ModelMap model, HttpSession session,
             @ModelAttribute("menu") MenuRequest menuRequest) {
+        if (Objects.equals(menuService.checkLogin(session), Boolean.FALSE)) {
+            model.addAttribute(ConstantManager.ERROR_POPUP, ConstantManager.NO_ACCEPTED_LOGIN);
+            return accController.initiate(model, session);
+        }
         List<Menu> m = menuService.checkLastInsert();
         if (m.get(0).getName().equals(menuRequest.getName()) && m.get(0).getParentId().equals(menuRequest.getParentId())) {
             return initiate(model, session);
@@ -67,6 +82,10 @@ public class MenuController {
     @RequestMapping(params = "insertSubMenu", method = RequestMethod.POST)
     public String insertSub(ModelMap model, HttpSession session,
             @ModelAttribute("menu") MenuRequest menuRequest) {
+        if (Objects.equals(menuService.checkLogin(session), Boolean.FALSE)) {
+            model.addAttribute(ConstantManager.ERROR_POPUP, ConstantManager.NO_ACCEPTED_LOGIN);
+            return accController.initiate(model, session);
+        }
         List<Menu> m = menuService.checkLastInsert();
         if (m.get(0).getName().equals(menuRequest.getName()) && m.get(0).getParentId().equals(menuRequest.getParentId())) {
             return getSubMenu(model, session, menuRequest.getParentId());
@@ -83,6 +102,10 @@ public class MenuController {
     @RequestMapping(params = "deleteMenu", method = RequestMethod.GET)
     public String delete(ModelMap model, HttpSession session,
             @RequestParam(MenuConstant.ID_PARAM) String id) {
+        if (Objects.equals(menuService.checkLogin(session), Boolean.FALSE)) {
+            model.addAttribute(ConstantManager.ERROR_POPUP, ConstantManager.NO_ACCEPTED_LOGIN);
+            return accController.initiate(model, session);
+        }
         if (Objects.equals(menuService.deleteMenu(id), Boolean.FALSE)) {
             model.put(ConstantManager.ERROR_POPUP, MenuConstant.DELETE_MENU_FAIL);
             return initiate(model, session);
@@ -94,6 +117,10 @@ public class MenuController {
     @RequestMapping(params = "deleteSubMenu", method = RequestMethod.GET)
     public String deleteSub(ModelMap model, HttpSession session,
             @RequestParam(MenuConstant.ID_PARAM) String id) {
+        if (Objects.equals(menuService.checkLogin(session), Boolean.FALSE)) {
+            model.addAttribute(ConstantManager.ERROR_POPUP, ConstantManager.NO_ACCEPTED_LOGIN);
+            return accController.initiate(model, session);
+        }
         Menu m = menuService.getMenuById(id);
         String fatherId = m.getParentId();
         if (Objects.equals(menuService.deleteMenu(id), Boolean.FALSE)) {
@@ -107,7 +134,10 @@ public class MenuController {
     @RequestMapping(params = "updateMenu", method = RequestMethod.POST)
     public String updateMenu(ModelMap model, HttpSession session,
             @ModelAttribute("menu") MenuRequest menu) {
-        
+        if (Objects.equals(menuService.checkLogin(session), Boolean.FALSE)) {
+            model.addAttribute(ConstantManager.ERROR_POPUP, ConstantManager.NO_ACCEPTED_LOGIN);
+            return accController.initiate(model, session);
+        }
         model.put(ConstantManager.ERROR_POPUP, MenuConstant.UPDATE_MENU_FAIL);
         return initiate(model, session);
     }
@@ -115,13 +145,16 @@ public class MenuController {
     @RequestMapping(params = "updateSubMenu", method = RequestMethod.POST)
     public String updateSubMenu(ModelMap model, HttpSession session,
             @ModelAttribute("sub") MenuRequest menu) {
-
+        if (Objects.equals(menuService.checkLogin(session), Boolean.FALSE)) {
+            model.addAttribute(ConstantManager.ERROR_POPUP, ConstantManager.NO_ACCEPTED_LOGIN);
+            return accController.initiate(model, session);
+        }
         menu.setCreatedBy((String) session.getAttribute("accountId"));
         if(Objects.equals(menuService.updateMenu(menu), Boolean.FALSE)){
             model.put(ConstantManager.ERROR_POPUP, "Update không thành công");
             return getSubMenu(model, session, menu.getParentId());
-        }
-        model.put(ConstantManager.ERROR_POPUP, MenuConstant.UPDATE_MENU_FAIL);
+        }     
+        model.put(ConstantManager.OK_POPUP, MenuConstant.UPDATE_MENU_OK);
         return getSubMenu(model, session, menu.getParentId());
     }
 
@@ -129,6 +162,10 @@ public class MenuController {
     public String lock(@RequestParam(MenuConstant.ID_PARAM) String id,
             @RequestParam(MenuConstant.STATUS_PARAM) Boolean status,
             ModelMap model, HttpSession session) {
+        if (Objects.equals(menuService.checkLogin(session), Boolean.FALSE)) {
+            model.addAttribute(ConstantManager.ERROR_POPUP, ConstantManager.NO_ACCEPTED_LOGIN);
+            return accController.initiate(model, session);
+        }
         if (Objects.equals(menuService.updateStatus(id, status), Boolean.FALSE)) {
             model.put(ConstantManager.ERROR_POPUP, "Thay đổi trạng thái không thành công");
             return initiate(model, session);
